{{cookiecutter.project_slug}}/requirements/base.txt
python-slugify==8.0.1
Pillow==10.0.0
python-decouple==3.8
{%- if cookiecutter.use_whitenoise == 'y' %}
whitenoise==6.5.0
{%- endif %}
argon2-cffi==21.3.0

# Django
# ------------------------------------------------------------------------------
django==4.2.4
django-cleanup==8.0.0
dj-database-url==2.0.0
{%- if cookiecutter.rest_framework == 'DRF' %}
# Django REST Framework
djangorestframework==3.14.0
django-cors-headers==4.2.0
# DRF-spectacular for api documentation
drf-spectacular==0.26.4
{%- elif cookiecutter.rest_framework == 'DNRF' %}
# Django Ninja REST framework
django-ninja==0.22.2
orjson==3.9.3
django-cors-headers==4.2.0
{%- endif %}
